changes in teardown scripts

+++ src/test/groovy/Specs/TearDown/TearDownProcessSpec.groovy
@@ -73,7 +73,7 @@ class TearDownProcessSpec extends BaseSpec {
         }
     }
 
-    @Test(priority = 2, description = "Delete all Automation Clinical Reports created during Automation",dependsOnMethods = "setupMethod")
+    @Test(priority = 2, description = "Delete all Automation Clinical Reports created during Automation", dependsOnMethods = "setupMethod")
     public void deleteClinicalReports() {
         String id;
         to LoginPage
@@ -85,32 +85,34 @@ class TearDownProcessSpec extends BaseSpec {
         at OmiciaHomePage
         getNamesOfAllWorkSpaces().each {
             workspace ->
-                try {
-                    id = getWorkspaceIdBasedOnName(workspace)
-                    switchWorkspace(workspace)
+                if (workspace.contains("ACMG_Automation_Workspace")) {
+                    try {
+                        id = getWorkspaceIdBasedOnName(workspace)
+                        switchWorkspace(workspace)
 
-                    at HeaderPage
-                    goToHomePage()
+                        at HeaderPage
+                        goToHomePage()
 
-                    at OmiciaHomePage
-                    openTab(CLINICAL_REPORTER);
+                        at OmiciaHomePage
+                        openTab(CLINICAL_REPORTER);
 
-                    at ClinicalReporterPage
-                    deleteAllClinicalReports()
+                        at ClinicalReporterPage
+                        deleteAllClinicalReports()
 
-                    at HeaderPage
-                    goToHomePage()
+                        at HeaderPage
+                        goToHomePage()
 
-                    at OmiciaHomePage
-                }
-                catch (Throwable throwable) {
-                    Assert.fail("Clinical Reporter Deletion Failed for WORKSPACE: " + workspace + " ,WORKSPACE_ID: " + id);
-                    throwable.printStackTrace()
+                        at OmiciaHomePage
+                    }
+                    catch (Throwable throwable) {
+                        Assert.fail("Clinical Reporter Deletion Failed for WORKSPACE: " + workspace + " ,WORKSPACE_ID: " + id);
+                        throwable.printStackTrace()
+                    }
                 }
         }
     }
 
-    @Test(priority = 3, description = "Delete all the Test Panels created during Automation",dependsOnMethods = "setupMethod")
+    @Test(priority = 3, description = "Delete all the Test Panels created during Automation", dependsOnMethods = "setupMethod")
     public void deletePanels() {
         String id;
         to LoginPage
@@ -122,34 +124,36 @@ class TearDownProcessSpec extends BaseSpec {
         at OmiciaHomePage
         getNamesOfAllWorkSpaces().each {
             workspace ->
-                try {
-                    if (workspace.contains("Automation")) {
-                        id = getWorkspaceIdBasedOnName(workspace)
-                        switchWorkspace(workspace)
+                if (workspace.contains("ACMG_Automation_Workspace")) {
+                    try {
+                        if (workspace.contains("Automation")) {
+                            id = getWorkspaceIdBasedOnName(workspace)
+                            switchWorkspace(workspace)
 
-                        at HeaderPage
-                        goToHomePage()
+                            at HeaderPage
+                            goToHomePage()
 
-                        at OmiciaHomePage
-                        openTab(PANEL_BUILDER);
+                            at OmiciaHomePage
+                            openTab(PANEL_BUILDER);
 
-                        at PanelBuilderPage
-                        deleteAllPanels()
+                            at PanelBuilderPage
+                            deleteAllPanels()
 
-                        at HeaderPage
-                        goToHomePage()
+                            at HeaderPage
+                            goToHomePage()
 
-                        at OmiciaHomePage
+                            at OmiciaHomePage
+                        }
                     }
-                }
-                catch (Throwable throwable) {
-                    Assert.fail("Panel Deletion Failed for WORKSPACE: " + workspace + " ,WORKSPACE_ID: " + id);
-                    throwable.printStackTrace()
+                    catch (Throwable throwable) {
+                        Assert.fail("Panel Deletion Failed for WORKSPACE: " + workspace + " ,WORKSPACE_ID: " + id);
+                        throwable.printStackTrace()
+                    }
                 }
         }
     }
 
-    @Test(priority = 4, description = "Delete all the Test Gene Sets created during Automation",dependsOnMethods = "setupMethod")
+    @Test(priority = 4, description = "Delete all the Test Gene Sets created during Automation", dependsOnMethods = "setupMethod")
     public void deleteAllGeneSets() {
         String id;
         to LoginPage
@@ -161,34 +165,36 @@ class TearDownProcessSpec extends BaseSpec {
         at OmiciaHomePage
         getNamesOfAllWorkSpaces().each {
             workspace ->
-                try {
-                    if (workspace.contains("Automation")) {
-                        id = getWorkspaceIdBasedOnName(workspace)
-                        switchWorkspace(workspace)
+                if (workspace.contains("ACMG_Automation_Workspace")) {
+                    try {
+                        if (workspace.contains("Automation")) {
+                            id = getWorkspaceIdBasedOnName(workspace)
+                            switchWorkspace(workspace)
 
-                        at HeaderPage
-                        goToHomePage()
+                            at HeaderPage
+                            goToHomePage()
 
-                        at OmiciaHomePage
-                        openTab(GENE_SETS);
+                            at OmiciaHomePage
+                            openTab(GENE_SETS);
 
-                        at GeneSetsPage
-                        deleteSets(MY_SET);
+                            at GeneSetsPage
+                            deleteSets(MY_SET);
 
-                        at HeaderPage
-                        goToHomePage()
+                            at HeaderPage
+                            goToHomePage()
 
-                        at OmiciaHomePage
+                            at OmiciaHomePage
+                        }
                     }
-                }
-                catch (Throwable throwable) {
-                    Assert.fail("Gene Sets Deletion Failed for WORKSPACE: " + workspace + " ,WORKSPACE_ID: " + id);
-                    throwable.printStackTrace()
+                    catch (Throwable throwable) {
+                        Assert.fail("Gene Sets Deletion Failed for WORKSPACE: " + workspace + " ,WORKSPACE_ID: " + id);
+                        throwable.printStackTrace()
+                    }
                 }
         }
     }
 
-    @Test(priority = 5, description = "Delete all the Test Filtering protocols created during Automation",dependsOnMethods = "setupMethod")
+    @Test(priority = 5, description = "Delete all the Test Filtering protocols created during Automation", dependsOnMethods = "setupMethod")
     public void deleteFilteringProtocols() {
         String id;
         to LoginPage
@@ -200,8 +206,8 @@ class TearDownProcessSpec extends BaseSpec {
         at OmiciaHomePage
         getNamesOfAllWorkSpaces().each {
             workspace ->
-                try {
-                    if (workspace.contains("Automation")) {
+                if (workspace.contains("ACMG_Automation_Workspace")) {
+                    try {
                         id = getWorkspaceIdBasedOnName(workspace)
                         switchWorkspace(workspace)
 
@@ -219,15 +225,15 @@ class TearDownProcessSpec extends BaseSpec {
 
                         at OmiciaHomePage
                     }
-                }
-                catch (Throwable throwable) {
-                    Assert.fail("Filtering Protocol Deletion Failed for WORKSPACE: " + workspace + " , WORKSPACE_ID: " + id);
-                    throwable.printStackTrace()
+                    catch (Throwable throwable) {
+                        Assert.fail("Filtering Protocol Deletion Failed for WORKSPACE: " + workspace + " , WORKSPACE_ID: " + id);
+                        throwable.printStackTrace()
+                    }
                 }
         }
     }
 
-    @Test(priority = 6, description = "Delete all the Test Projects created during Automation",dependsOnMethods = "setupMethod")
+    @Test(priority = 6, description = "Delete all the Test Projects created during Automation", dependsOnMethods = "setupMethod")
     public void deleteAllProjects() {
         String id;
         to LoginPage
@@ -239,8 +245,8 @@ class TearDownProcessSpec extends BaseSpec {
         at OmiciaHomePage
         getNamesOfAllWorkSpaces().each {
             workspace ->
-                try {
-                    if (workspace.contains("Automation")) {
+                if (workspace.contains("ACMG_Automation_Workspace")) {
+                    try {
                         id = getWorkspaceIdBasedOnName(workspace)
                         switchWorkspace(workspace)
 
@@ -266,15 +272,15 @@ class TearDownProcessSpec extends BaseSpec {
 
                         at OmiciaHomePage
                     }
-                }
-                catch (Throwable throwable) {
-                    Assert.fail("Project Deletion Failed for WORKSPACE: " + workspace + ", WORKSPACE_ID: " + id);
-                    throwable.printStackTrace()
+                    catch (Throwable throwable) {
+                        Assert.fail("Project Deletion Failed for WORKSPACE: " + workspace + ", WORKSPACE_ID: " + id);
+                        throwable.printStackTrace()
+                    }
                 }
         }
     }
 
-    @Test(priority = 7, description = "Remove test user from the Automation Workspaces",dependsOnMethods = "setupMethod")
+    @Test(priority = 7, description = "Remove test user from the Automation Workspaces", dependsOnMethods = "setupMethod")
     public void deleteUsersFromWorkspaces() {
         String id;
         to LoginPage
